pyproject.toml
[project]
authors = [{ name = "spoddub", email = "spoddub98@gmail.com" }]
name = "hexlet-code"
version = "0.1.0"
description = "CLI utility that compares two configuration files and shows a difference."
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pyyaml"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.1",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pyyaml>=6.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]
